regulatory: extract 24hr/8hr EPA metrics from surfplots into util/regulatory_tools
