+++ melodies_monet/plots/surfplots.py
@@ -34,58 +34,6 @@ except ImportError:
     windrose = None
     WindroseAxes = None
 
-def make_24hr_regulatory(df, col=None):
-    """Calculates 24-hour averages
-    
-    Parameters
-    ----------
-    df : dataframe
-        Model/obs pair of hourly data
-    col : str
-        Column label of observation variable to apply the calculation 
-    Returns
-    -------
-    dataframe
-        dataframe with applied calculation
-        
-    """
-    #return calc_24hr_ave(df, col)
-    return calc_24hr_ave_v1(df, col)
-
-def calc_24hr_ave_v1(df, col=None):
-    df.index = df.time_local
-    # select sites with nobs >=18, 75% completeness
-    df_24hr_ave = (df.groupby("siteid")[col].resample("D").sum(min_count=18, numeric_only=True)/df.groupby("siteid")[col].resample("D").count()).reset_index().dropna()
-    df = df.reset_index(drop=True)
-    return df.merge(df_24hr_ave, on=["siteid", "time_local"])
-
-def make_8hr_regulatory(df, col=None):
-    """Calculates 8-hour rolling average daily
-    
-    Parameters
-    ----------
-    df : dataframe
-        Model/obs pair of hourly data
-    col : str
-        Column label of observation variable to apply the calculation 
-    Returns
-    -------
-    dataframe
-        dataframe with applied calculation
-        
-    """
-    #return calc_8hr_rolling_max(df, col, window=8)
-    return calc_8hr_rolling_max_v1(df, col, window=8)
-
-def calc_8hr_rolling_max_v1(df, col=None, window=None):
-    df.index = df.time_local
-    df_rolling = df.groupby("siteid")[col].rolling(window,min_periods=6,center=True, win_type="boxcar").mean(numeric_only=True).reset_index().dropna()
-    # JianHe: select sites with nobs >=18, 75% completeness based on EPA
-    df_rolling.index = df_rolling.time_local
-    df_rolling_max = df_rolling.groupby("siteid").resample("D").max(min_count=18, numeric_only=True).reset_index().dropna()
-    df = df.reset_index(drop=True)
-    return df.merge(df_rolling_max, on=["siteid", "time_local"])
-
 def calc_default_colors(p_index):
     """List of default colors, lines, and markers to use if user does not 
     specify them in the input yaml file.
